Address final review blockers
