+++ test/helpers/context-fixtures.ts
@@ -1,0 +1,184 @@
+import { vi } from "vitest";
+import type { AssistantMessage, AssistantMessageEvent, Context } from "@earendil-works/pi-ai";
+import {
+	AuthStorage,
+	ModelRegistry,
+	type BuildSystemPromptOptions,
+	type ExtensionCommandContext,
+	type ExtensionContext,
+} from "@earendil-works/pi-coding-agent";
+import { makeModel } from "./model-fixtures.js";
+import type { ExtensionCommandContextOverrides, ExtensionContextOverrides } from "./pi-harness-types.js";
+
+let sharedTestModelRegistry: ModelRegistry | undefined;
+
+function getSharedTestModelRegistry(): ModelRegistry {
+	sharedTestModelRegistry ??= ModelRegistry.inMemory(AuthStorage.inMemory());
+	return sharedTestModelRegistry;
+}
+
+export function createDefaultSystemPromptOptions(cwd: string): BuildSystemPromptOptions {
+	return {
+		cwd,
+		selectedTools: ["read", "bash", "edit", "write"],
+	};
+}
+
+function createMinimalSessionManager(overrides: Partial<ExtensionContext["sessionManager"]> = {}): ExtensionContext["sessionManager"] {
+	return {
+		getCwd: vi.fn(() => process.cwd()),
+		getSessionDir: vi.fn(() => ""),
+		getSessionId: vi.fn(() => "test-session"),
+		getSessionFile: vi.fn(() => undefined),
+		getLeafId: vi.fn(() => null),
+		getLeafEntry: vi.fn(() => undefined),
+		getEntry: vi.fn(() => undefined),
+		getLabel: vi.fn(() => undefined),
+		getBranch: vi.fn(() => []),
+		getHeader: vi.fn(() => null),
+		getEntries: vi.fn(() => []),
+		getTree: vi.fn(() => []),
+		getSessionName: vi.fn(() => undefined),
+		...overrides,
+	};
+}
+
+function createMinimalExtensionUi(): ExtensionContext["ui"] {
+	return {
+		select: vi.fn(async () => undefined),
+		confirm: vi.fn(async () => false),
+		input: vi.fn(async () => undefined),
+		notify: vi.fn(),
+		onTerminalInput: vi.fn(() => () => {}),
+		setStatus: vi.fn(),
+		setWorkingMessage: vi.fn(),
+		setWorkingVisible: vi.fn(),
+		setWorkingIndicator: vi.fn(),
+		setHiddenThinkingLabel: vi.fn(),
+		setWidget: vi.fn(),
+		setFooter: vi.fn(),
+		setHeader: vi.fn(),
+		setTitle: vi.fn(),
+		custom: vi.fn(async () => undefined as never),
+		pasteToEditor: vi.fn(),
+		setEditorText: vi.fn(),
+		getEditorText: vi.fn(() => ""),
+		editor: vi.fn(async () => undefined),
+		addAutocompleteProvider: vi.fn(),
+		setEditorComponent: vi.fn(),
+		getEditorComponent: vi.fn(() => undefined),
+		theme: {} as ExtensionContext["ui"]["theme"],
+		getAllThemes: vi.fn(() => []),
+		getTheme: vi.fn(() => undefined),
+		setTheme: vi.fn(() => ({ success: true })),
+		getToolsExpanded: vi.fn(() => false),
+		setToolsExpanded: vi.fn(),
+	} satisfies ExtensionContext["ui"];
+}
+
+function createMinimalExtensionContextInternal(overrides: ExtensionContextOverrides = {}): ExtensionContext {
+	const cwd = overrides.cwd ?? process.cwd();
+	const base: ExtensionContext = {
+		ui: createMinimalExtensionUi(),
+		hasUI: true,
+		cwd,
+		sessionManager: createMinimalSessionManager(),
+		modelRegistry: getSharedTestModelRegistry(),
+		model: makeModel("composer-2.5"),
+		isIdle: vi.fn(() => true),
+		signal: undefined,
+		abort: vi.fn(),
+		hasPendingMessages: vi.fn(() => false),
+		shutdown: vi.fn(),
+		getContextUsage: vi.fn(() => undefined),
+		compact: vi.fn(),
+		getSystemPrompt: vi.fn(() => ""),
+	};
+	return {
+		...base,
+		...overrides,
+		ui: {
+			...base.ui,
+			...overrides.ui,
+		},
+		sessionManager: {
+			...base.sessionManager,
+			...overrides.sessionManager,
+		},
+	};
+}
+
+function createMinimalExtensionCommandContextInternal(
+	overrides: ExtensionCommandContextOverrides = {},
+): ExtensionCommandContext {
+	const base = createMinimalExtensionContextInternal(overrides) as ExtensionCommandContext;
+	return {
+		...base,
+		...overrides,
+		waitForIdle: overrides.waitForIdle ?? vi.fn(async () => undefined),
+		newSession: overrides.newSession ?? vi.fn(async () => ({ cancelled: false })),
+		fork: overrides.fork ?? vi.fn(async () => ({ cancelled: false })),
+		navigateTree: overrides.navigateTree ?? vi.fn(async () => ({ cancelled: false })),
+		switchSession: overrides.switchSession ?? vi.fn(async () => ({ cancelled: false })),
+		reload: overrides.reload ?? vi.fn(async () => undefined),
+		ui: {
+			...base.ui,
+			...overrides.ui,
+		},
+		sessionManager: {
+			...base.sessionManager,
+			...overrides.sessionManager,
+		},
+	};
+}
+
+export function createExtensionTestContext(ctxOverrides: ExtensionContextOverrides = {}): ExtensionContext {
+	return createMinimalExtensionContextInternal(ctxOverrides);
+}
+
+export function createExtensionCommandContext(
+	ctxOverrides: ExtensionCommandContextOverrides = {},
+): ExtensionCommandContext {
+	return createMinimalExtensionCommandContextInternal(ctxOverrides);
+}
+
+export function makeContext(messages: Context["messages"] = [{ role: "user", content: "Hello", timestamp: 1 }]): Context {
+	return {
+		systemPrompt: "Be helpful.",
+		messages,
+	};
+}
+
+export function makeAssistantMessage(text = "Done", timestamp = 2): AssistantMessage {
+	return {
+		role: "assistant",
+		content: [{ type: "text", text }],
+		api: "cursor-sdk",
+		provider: "cursor",
+		model: "test-model",
+		usage: {
+			input: 0,
+			output: 0,
+			cacheRead: 0,
+			cacheWrite: 0,
+			totalTokens: 0,
+			cost: { input: 0, output: 0, cacheRead: 0, cacheWrite: 0, total: 0 },
+		},
+		stopReason: "stop",
+		timestamp,
+	};
+}
+
+export async function collectEvents<TEvent>(stream: AsyncIterable<TEvent>): Promise<TEvent[]> {
+	const events: TEvent[] = [];
+	for await (const event of stream) {
+		events.push(event);
+	}
+	return events;
+}
+
+export async function collectAssistantEvents(
+	stream: AsyncIterable<AssistantMessageEvent>,
+): Promise<AssistantMessageEvent[]> {
+	return collectEvents(stream);
+}
